+++ History.md
@@ -1,4 +1,9 @@
+Next
+===================
+
+  * `purify` now accepts an options object, instead of prefix and postfix strings
+
 v0.1.0 / 2014-10-03
-==================
+===================
 
   * Initial release